Ajusta la creacion de la base de datos en docker, ajusta CRUD ordenes de compra, crea CRUD completo para Detalles Ordenes de compras

+++ src/main/java/com/hersac/herp/modulos/comercial/ordenesCompra/entidades/OrdenCompraEntity.java
@@ -21,6 +21,6 @@ public class OrdenCompraEntity {
 
     private Date fecha;
     private Boolean estaActiva;
-    private Integer total;
+    private Double total;
     private Long proveedorId;
 }
